CLN: remove keep_null_freq in ops (#30221)
