+++ pandas/tests/arithmetic/test_timedelta64.py
@@ -1046,6 +1046,58 @@ class TestTimedeltaArraylikeAddSubOps:
         with pytest.raises(TypeError):
             vector - tdser
 
+    # TODO: parameterize over box and de-duplicate
+    def test_tdi_add_sub_int(self, one):
+        # Variants of `one` for #19012, deprecated GH#22535
+        rng = timedelta_range("1 days 09:00:00", freq="H", periods=10)
+        msg = "Addition/subtraction of integers"
+
+        with pytest.raises(TypeError, match=msg):
+            rng + one
+        with pytest.raises(TypeError, match=msg):
+            rng += one
+        with pytest.raises(TypeError, match=msg):
+            rng - one
+        with pytest.raises(TypeError, match=msg):
+            rng -= one
+
+    # TODO: parameterize over box and de-duplicate
+    @pytest.mark.parametrize("box", [np.array, pd.Index])
+    def test_tdi_add_sub_integer_array(self, box):
+        # GH#19959, deprecated GH#22535
+        rng = timedelta_range("1 days 09:00:00", freq="H", periods=3)
+        other = box([4, 3, 2])
+        msg = "Addition/subtraction of integers and integer-arrays"
+
+        with pytest.raises(TypeError, match=msg):
+            rng + other
+
+        with pytest.raises(TypeError, match=msg):
+            other + rng
+
+        with pytest.raises(TypeError, match=msg):
+            rng - other
+
+        with pytest.raises(TypeError, match=msg):
+            other - rng
+
+    # TODO: parameterize over box and de-duplicate
+    @pytest.mark.parametrize("box", [np.array, pd.Index])
+    def test_tdi_addsub_integer_array_no_freq(self, box):
+        # GH#19959
+        tdi = TimedeltaIndex(["1 Day", "NaT", "3 Hours"])
+        other = box([14, -1, 16])
+        msg = "Addition/subtraction of integers"
+
+        with pytest.raises(TypeError, match=msg):
+            tdi + other
+        with pytest.raises(TypeError, match=msg):
+            other + tdi
+        with pytest.raises(TypeError, match=msg):
+            tdi - other
+        with pytest.raises(TypeError, match=msg):
+            other - tdi
+
     # ------------------------------------------------------------------
     # Operations with timedelta-like others
