+++ pandas/tests/indexes/timedeltas/test_arithmetic.py
@@ -98,61 +98,6 @@ class TestTimedeltaIndexArithmetic:
             tdi.shift(2)
 
     # -------------------------------------------------------------
-    # Binary operations TimedeltaIndex and integer
-
-    def test_tdi_add_sub_int(self, one):
-        # Variants of `one` for #19012, deprecated GH#22535
-        rng = timedelta_range("1 days 09:00:00", freq="H", periods=10)
-        msg = "Addition/subtraction of integers"
-
-        with pytest.raises(TypeError, match=msg):
-            rng + one
-        with pytest.raises(TypeError, match=msg):
-            rng += one
-        with pytest.raises(TypeError, match=msg):
-            rng - one
-        with pytest.raises(TypeError, match=msg):
-            rng -= one
-
-    # -------------------------------------------------------------
-    # __add__/__sub__ with integer arrays
-
-    @pytest.mark.parametrize("box", [np.array, pd.Index])
-    def test_tdi_add_sub_integer_array(self, box):
-        # GH#19959, deprecated GH#22535
-        rng = timedelta_range("1 days 09:00:00", freq="H", periods=3)
-        other = box([4, 3, 2])
-        msg = "Addition/subtraction of integers and integer-arrays"
-
-        with pytest.raises(TypeError, match=msg):
-            rng + other
-
-        with pytest.raises(TypeError, match=msg):
-            other + rng
-
-        with pytest.raises(TypeError, match=msg):
-            rng - other
-
-        with pytest.raises(TypeError, match=msg):
-            other - rng
-
-    @pytest.mark.parametrize("box", [np.array, pd.Index])
-    def test_tdi_addsub_integer_array_no_freq(self, box):
-        # GH#19959
-        tdi = TimedeltaIndex(["1 Day", "NaT", "3 Hours"])
-        other = box([14, -1, 16])
-        msg = "Addition/subtraction of integers"
-
-        with pytest.raises(TypeError, match=msg):
-            tdi + other
-        with pytest.raises(TypeError, match=msg):
-            other + tdi
-        with pytest.raises(TypeError, match=msg):
-            tdi - other
-        with pytest.raises(TypeError, match=msg):
-            other - tdi
-
-    # -------------------------------------------------------------
     # Binary operations TimedeltaIndex and timedelta-like
     # Note: add and sub are tested in tests.test_arithmetic, in-place
     #  tests are kept here because their behavior is Index-specific
